+++ packages/basics/aprilio/c_src/lua_string.h
@@ -65,7 +65,7 @@ namespace AprilIO {
 
     virtual bool isOpened() const;
     virtual void close();
-    virtual off_t seek(int whence = SEEK_CUR, int offset = 0);
+    virtual off_t seek(int whence = SEEK_CUR, long offset = 0);
     virtual void flush();
     virtual int setvbuf(int mode, size_t size);
     virtual bool hasError() const;
@@ -114,7 +114,7 @@ namespace AprilIO {
     
     virtual bool isOpened() const;
     virtual void close();
-    virtual off_t seek(int whence = SEEK_CUR, int offset = 0);
+    virtual off_t seek(int whence = SEEK_CUR, long offset = 0);
     virtual int setvbuf(int mode, size_t size);
     virtual bool hasError() const;
     virtual const char *getErrorMsg() const;
